requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
openai==1.10.0
azure-storage-blob==12.19.0
azure-ai-formrecognizer==3.3.2
azure-search-documents==11.4.0
azure-core==1.30.0
numpy